add lr_scheduler
